[P4ADEV-2047] Code review
