+++ src/main/java/it/gov/pagopa/pu/bff/connector/process_executions/ExportFileServiceImpl.java
@@ -1,6 +1,7 @@
 package it.gov.pagopa.pu.bff.connector.process_executions;
 
 import it.gov.pagopa.pu.bff.connector.process_executions.client.ExportFileClient;
+import it.gov.pagopa.pu.bff.connector.process_executions.client.ExportFileSearchClient;
 import it.gov.pagopa.pu.bff.dto.ExportFileFiltersDTO;
 import it.gov.pagopa.pu.processexecutions.dto.generated.ExportFileRequestDTO;
 import it.gov.pagopa.pu.processexecutions.dto.generated.PagedModelExportFile;
@@ -11,16 +12,18 @@ import org.springframework.stereotype.Service;
 public class ExportFileServiceImpl implements ExportFileService {
 
   private final ExportFileClient client;
+  private final ExportFileSearchClient searchClient;
 
-  public ExportFileServiceImpl(ExportFileClient client) {
+  public ExportFileServiceImpl(ExportFileClient client, ExportFileSearchClient searchClient) {
     this.client = client;
+    this.searchClient = searchClient;
   }
 
   @Override
   public PagedModelExportFile getExportFiles(
     ExportFileFiltersDTO exportFileFilters,
     String operatorExternalId, Pageable pageable, String accessToken) {
-    return client.getExportFiles(exportFileFilters, operatorExternalId,
+    return searchClient.getExportFiles(exportFileFilters, operatorExternalId,
       pageable, accessToken);
   }
